+++ libs/newsletters-data-client/src/lib/transformDataToLegacyNewsletter.ts
